Revert "Fix tests for ARM like in commit 9bc9d70 for x86_64"
This reverts commit 1a714389321bcaf6448367956b666e81938efdfe.

+++ tests/output/c-backend/array-arm.c
@@ -4,5 +4,5 @@ void func(int arg)
   int a;
 #line 2 "/home/novikov/work/cif/tests/input/c-backend/array.c"
   int *buf;
-  a = * ( buf + arg );
+  a = * ( buf + ( unsigned int ) arg );
 }
